Example of common pointer usage mistake
Pointers in C are meant to reference the memory address of a certain variable. However, in certain scenarios, the reference may get stuck in a different stack generated by some other function in your code and it may never refer back to the address that you originally sent. This simple C program aims at recreating that.